refactor SubscriptionBean traffic stats storage

+++ app/src/main/java/io/nekohasekai/sagernet/bg/SubscriptionUpdater.kt
@@ -85,7 +85,7 @@ object SubscriptionUpdater {
             if (subscriptions.isNotEmpty()) for (profile in subscriptions) {
                 val subscription = profile.subscription!!
 
-                if (((System.currentTimeMillis() / 1000).toInt() - subscription.lastUpdated) < subscription.autoUpdateDelay * 60) {
+                if ((System.currentTimeMillis() / 1000 - subscription.lastUpdated) < subscription.autoUpdateDelay * 60) {
                     continue
                 }
